requirements.txt
llama-cpp-python==0.1.78
pandas
huggingface_hub
openai
scriptify